// src/web/src/riva/projects/riva-shared/src/lib/layout/enums/layout-menu-body-class.enum.ts
export enum LayoutMenuBodyClass {
    FixedNavbar = 'fixed-navbar',
    HorizontalLayout = 'horizontal-layout',
    HorizontalMenu = 'horizontal-menu',
    HorizontalMenuPadding = 'horizontal-menu-padding',
    MenuExpanded = 'menu-expanded',
    MenuHide = 'menu-hide',
    MenuOpen = 'menu-open',
    NavCollapsed = 'nav-collapsed',
    NavbarScrolled = 'navbar-scrolled',
    OverflowHidden = 'overflow-hidden',
    PageScrolled = 'page-scrolled',
    VerticalLayout = 'vertical-layout',
    VerticalMenu = 'vertical-menu',
    VerticalOverlayMenu = 'vertical-overlay-menu'
}